Move broken map into its own source file
